SERVER

+++ src/main/java/com/neighborcharger/capstoneproject/DTO/ChargingCarDTO.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 public class ChargingCarDTO {
     String runtime;        // 얼마나 충전했는가
     double usingElectric; // 사용한 전기량?
-    double Cost; // 비용
+    int Cost; // 비용
 }
